mycket saker

+++ src/main/java/com/example/pokeblitz/Repositories/BattleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BattleRepository extends CrudRepository<Battle, Integer> {
+public interface BattleRepository extends CrudRepository<Battle, Long> {
 
 }
