+++ src/main/java/com/example/pokeblitz/Repositories/PlayerRepository.java
@@ -3,6 +3,6 @@ package com.example.pokeblitz.Repositories;
 import com.example.pokeblitz.Classes.Player;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlayerRepository extends CrudRepository<Player, Integer> {
+public interface PlayerRepository extends CrudRepository<Player, Long> {
     Player findByusername(String username);
 }
